FINAL FINALOV

+++ forum.php
@@ -1,6 +1,7 @@
 <?php
 session_start();
 require_once 'classes/ForumSection.php';
+require_once 'classes/Database.php';
 
 if (!isset($_SESSION['user'])) {
     echo "<div style='color: red; text-align: center;'>You must be logged in to participate in the forum. <a href='index.php'>Login here</a>.</div>";
@@ -17,7 +18,6 @@ $result = $stmt->get_result();
 $userData = $result->fetch_assoc();
 $loggedUserId = $userData['id'];
 
-
 if ($_SERVER['REQUEST_METHOD'] === 'POST' && isset($_POST['action']) && $_POST['action'] === 'create_section') {
     $name = trim($_POST['name']);
     $description = trim($_POST['description']);
@@ -30,7 +30,6 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST' && isset($_POST['action']) && $_POST['
         echo "<p style='color:red;'>Section name cannot be empty.</p>";
     }
 }
-
 
 if (isset($_GET['delete_section'])) {
     $sectionIdToDelete = (int)$_GET['delete_section'];
@@ -46,7 +45,6 @@ if (isset($_GET['delete_section'])) {
     }
 }
 
-
 $stmt = $conn->prepare("SELECT fs.*, u.username, u.profile_picture 
                         FROM forum_sections fs
                         JOIN users u ON fs.user_id = u.id
@@ -55,7 +53,7 @@ $stmt->execute();
 $result = $stmt->get_result();
 $sections = $result->fetch_all(MYSQLI_ASSOC);
 
-include 'kokosy/header.php';
+include 'partials/header.php';
 ?>
 
 <div id="forum-content">
@@ -74,33 +72,37 @@ include 'kokosy/header.php';
 
     <h2>Forum Sections</h2>
 
-    <?php foreach ($sections as $section): ?>
-        <div class='forum-section'>
-            <div style="display: flex; align-items: center;">
-                <?php if (!empty($section['profile_picture'])): ?>
-                    <img src="<?php echo htmlspecialchars($section['profile_picture']); ?>" alt="Profile Pic" width="40" height="40" style="border-radius:50%; margin-right:10px;">
-                <?php else: ?>
-                    <img src="images/profile_pics/default.png" alt="Profile Pic" width="40" height="40" style="border-radius:50%; margin-right:10px;">
-                <?php endif; ?>
+    <?php if (count($sections) > 0): ?>
+        <?php foreach ($sections as $section): ?>
+            <div class='forum-section'>
+                <div style="display: flex; align-items: center;">
+                    <?php if (!empty($section['profile_picture'])): ?>
+                        <img src="<?php echo htmlspecialchars($section['profile_picture']); ?>" alt="Profile Pic" width="40" height="40" style="border-radius:50%; margin-right:10px;">
+                    <?php else: ?>
+                        <img src="images/profile_pics/default.jpg" alt="Profile Pic" width="40" height="40" style="border-radius:50%; margin-right:10px;">
+                    <?php endif; ?>
 
-                <h3 style="margin: 0;">
-                    <a href='forum_section.php?id=<?php echo $section['id']; ?>'>
-                        <?php echo htmlspecialchars($section['name']); ?>
-                    </a>
-                </h3>
-            </div>
-            <p><?php echo nl2br(htmlspecialchars($section['description'])); ?></p>
-            <p><em>Created by <?php echo htmlspecialchars($section['username']); ?></em></p>
-
-            <?php if ($section['user_id'] == $loggedUserId): ?>
-                <div>
-                    <a href='forum.php?delete_section=<?php echo $section['id']; ?>' style='color:red;'>Delete Section</a> |
-                    <a href='edit_section.php?section_id=<?php echo $section['id']; ?>'>Edit Section</a>
+                    <h3 style="margin: 0;">
+                        <a href='forum_section.php?id=<?php echo $section['id']; ?>'>
+                            <?php echo htmlspecialchars($section['name']); ?>
+                        </a>
+                    </h3>
                 </div>
-            <?php endif; ?>
-        </div>
-    <?php endforeach; ?>
+                <p><?php echo nl2br(htmlspecialchars($section['description'])); ?></p>
+                <p><em>Created by <?php echo htmlspecialchars($section['username']); ?></em></p>
+
+                <?php if ($section['user_id'] == $loggedUserId): ?>
+                    <div>
+                        <a href='forum.php?delete_section=<?php echo $section['id']; ?>' style='color:red;'>Delete Section</a> |
+                        <a href='forum_actions/edit_section.php?section_id=<?php echo $section['id']; ?>'>Edit Section</a>
+                    </div>
+                <?php endif; ?>
+            </div>
+        <?php endforeach; ?>
+    <?php else: ?>
+        <p>No forum sections yet.</p>
+    <?php endif; ?>
 
 </div>
 
-<?php include 'kokosy/footer.php'; ?>
+<?php include 'partials/footer.php'; ?>
